LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-stretch-properties-001-expected.txt


FAIL Stretchy vertical line assert_approx_equals: Prefix <mo> element expected 150 +/- 1 but got 25
|
|
|
|
